main_src_code/backend/requirement.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.3.2
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
threadpoolctl==2.2.0
tinydb==4.5.1
Werkzeug==2.0.1
